conversion to float64, including weight functions

+++ data_preparation.py
@@ -43,8 +43,8 @@ def process_data(waterlevelFrame, ratesFrame, maxGaps):
     ## convert data to expected data types
     timeseries = waterlevelFrame["Time"] - waterlevelFrame["Time"].iloc[0]
     timeseries = np.array(timeseries) / pd.Timedelta(1, unit='d')
-    waterlevel = np.array(waterlevelFrame["Waterlevel"])
-    rates = np.array(ratesFrame["Gesamt"])
+    waterlevel = np.array(waterlevelFrame["Waterlevel"], dtype = "float")
+    rates = np.array(ratesFrame["Gesamt"], dtype = "float")
     return timeseries, waterlevel, rates, isInterWl, isInterRates
 
 def cut_time_range(waterlevelFrame, ratesFrame):
